Interrupt 삭제

+++ Raspberry/process.cpp
@@ -6,8 +6,6 @@
 #include <iostream>
 
 #define HUMAN 1			// Human detect sensor wPi Pin 1, BGM 18
-#define LED_G 2			// LED wPi pin 2, BGM 2 7
-#define LED_R 3
 
 using namespace std;
 unsigned char exit_flag = 0;
@@ -18,22 +16,12 @@ PROCESS::PROCESS()
     system("sudo motion");			// Motion 실행
     cout << "CCTV Enable" << endl;
 
-    pinMode(LED_G, OUTPUT);
-    pinMode(LED_R, OUTPUT);			// LED 핀 초기화
     pinMode(HUMAN, INPUT);			// 인체감지센서 초기화
-
-    digitalWrite(LED_R, LOW);
-    digitalWrite(LED_G, HIGH);
 
     signal(SIGINT, signal_handler);
     signal(SIGTERM, signal_handler);		// 강제 종료 시 처리되는 인터럽트 처리
 
     atexit(call_exitfunc);			// 시스템 종료 시 호출되는 함수 선언
-    if (wiringPiISR(HUMAN, INT_EDGE_RISING, &humanInterrupt) < 0)
-    {
-        exit(1);
-    }						// 인체감지센서 인터럽트 선언 감지 시 humanInterrput 함수 실행
-    else std::cout << "Interrupt enabled\n";
 }
 void PROCESS::processCycle()
 {
@@ -45,8 +33,8 @@ void PROCESS::processCycle()
 void PROCESS::printData()
 {
     cout << "============================" << endl;
-    cout << " temp = " << dht.getTemp() << "hum = " << dht.getHum() << endl;
-    cout << "PM1 = " << pms.getPM() << endl;
+    cout << "temp = " << dht.getTemp() << " hum = " << dht.getHum() << endl;
+    cout << "PM10 = " << pms.getPM() << endl;
     cout << "============================" << endl;
 }
 void PROCESS::putData(int* t, int* h, int* d)
@@ -55,18 +43,11 @@ void PROCESS::putData(int* t, int* h, int* d)
     *t = dht.getTemp();
     *d = pms.getPM();
 }
-void humanInterrupt()
-{
-    cout << "Human detect" << endl;
-}
-
 void signal_handler(int signo)
 {
     cout << "call signal handler" << endl;
     system("sudo killall -15 motion");
     exit_flag = 1;
-    digitalWrite(LED_G, LOW);
-    digitalWrite(LED_R, HIGH);
 
     exit(0);
 }
@@ -74,11 +55,6 @@ void signal_handler(int signo)
 void call_exitfunc()
 {
     cout << "call_ateixt func" << endl;
-    if (exit_flag == 0)
-    {
-        system("sudo killall -15 motion");
-        digitalWrite(LED_G, LOW);
-        digitalWrite(LED_R, HIGH);
-    }
+    if (exit_flag == 0)     system("sudo killall -15 motion");
     cout << "CCTV disabled" << endl;
 } // 시스템 종료 시 호출되는 함수
